:ok_hand: [#77] feat: Process PR Feedback
